feat(StringCalculator): 클래스 분리
